A first attempt at evaluation by separation.
Basically, we looked at all the code we created so far and decided it was crap.
So we came up with another algorithm that just assumes we can decode chromosomes
into symbols and goes from there. Rather than error correcting the chromosome,
we instead separate it into digits and operators, and then construct a legal
expression. Essentially, the chromosome is now an anagram for a legal
expression, and we construct the first one we find.